+++ CHANGELOG.md
@@ -10,6 +10,8 @@
  - Any type of `Response` can be converted into futures 0.1 `Future`, or
    `std::future::Future`.
  - Add implementations of `Response` for pointer types.
+ - Add `ResponseExt::with_backoff`, which provides retry behavior with
+   exponential backoff.
 
 ### Bug fixes